Fix test failure.
The way zc.buildout.testing.install works is it takes eggs from the
outer buildout and copies them into the inner one.  Therefore we must
ensure the egg exists in the outer one, or not try to copy unneeded eggs.

And since zope.browserpage is apparently no longer needed, drop it and
(almost) all of its dependencies I'd added in r129557.